Add user role

+++ app.py
@@ -17,8 +17,8 @@ app.config["SQLALCHEMY_DATABASE_URI"] = "mysql+pymysql://root:admin123@127.0.0.1
 login_manager = LoginManager()
 db.init_app(app)
 login_manager.init_app(app)
-login_manager.login_view = "login"
 # view Login
+login_manager.login_view = "login"
 
 
 @login_manager.user_loader
@@ -56,7 +56,7 @@ def create_user():
     password = data.get("password")
 
     if username and password:
-        user = User(username=username, password=password)
+        user = User(username=username, password=password, role="user")
         db.session.add(user)
         db.session.commit()        
         return jsonify({ "message": "Usuário cadastrado com sucesso" })
@@ -81,6 +81,9 @@ def update_user(id_user):
     data = request.json
     user = User.query.get(id_user)
 
+    if id_user != current_user.id and current_user.role == "user":
+        return jsonify({ "message" : "Operação não permitida" }), 403
+
     new_password = data.get("password")
     if user and new_password:
         user.password = new_password
@@ -96,8 +99,8 @@ def update_user(id_user):
 def delete_user(id_user):
     user = User.query.get(id_user)
 
-    if id_user == current_user.id:
-        return jsonify({ "message": "Deleção não permitida" }), 403
+    if id_user == current_user.id or current_user.role != "admin":
+        return jsonify({ "message" : "Operação não permitida" }), 403
 
     if user:
         db.session.delete(user)
